[FOLD] Move TxQ to app/misc, devirtualize, and levelize. * Also split apply.h/cpp into applySteps.h/cpp cleanly. * No logic changes to TxQ, but a lot of code moved. * Rules::operator= does a short-circuit compare for equivalent pointers before doing the deep compare.
