+++ src/ripple/app/misc/TxQ.h
@@ -20,16 +20,104 @@
 #ifndef RIPPLE_TXQ_H_INCLUDED
 #define RIPPLE_TXQ_H_INCLUDED
 
+#include <ripple/app/tx/applySteps.h>
 #include <ripple/ledger/OpenView.h>
 #include <ripple/ledger/ApplyView.h>
 #include <ripple/core/Config.h>
 #include <ripple/core/LoadFeeTrack.h>
 #include <ripple/protocol/TER.h>
 #include <ripple/protocol/STTx.h>
+#include <boost/intrusive/set.hpp>
 
 namespace ripple {
 
 class Application;
+
+namespace detail {
+
+class FeeMetrics
+{
+private:
+    // Fee escalation
+
+    // Limit of the txnsExpected value after a
+    // time leap.
+    std::size_t const targetTxnCount_;
+    // Minimum value of txnsExpected.
+    std::size_t minimumTxnCount_;
+    // Number of transactions expected per ledger.
+    // One more than this value will be accepted
+    // before escalation kicks in.
+    std::size_t txnsExpected_;
+    // Minimum value of escalationMultiplier.
+    std::uint32_t const minimumMultiplier_;
+    // Based on the median fee of the LCL. Used
+    // when fee escalation kicks in.
+    std::uint32_t escalationMultiplier_;
+    beast::Journal j_;
+
+    std::mutex mutable lock_;
+
+public:
+    static const std::uint64_t baseLevel = 256;
+
+public:
+    FeeMetrics(bool standAlone, beast::Journal j)
+        : targetTxnCount_(50)
+        , minimumTxnCount_(standAlone ? 1000 : 5)
+        , txnsExpected_(minimumTxnCount_)
+        , minimumMultiplier_(500)
+        , escalationMultiplier_(minimumMultiplier_)
+        , j_(j)
+    {
+    }
+
+    /**
+    Updates fee metrics based on the transactions in the ReadView
+    for use in fee escalation calculations.
+
+    @param view View of the LCL that was just closed or received.
+    @param timeLeap Indicates that rippled is under load so fees
+    should grow faster.
+    */
+    std::size_t
+    updateFeeMetrics(Application& app,
+        ReadView const& view, bool timeLeap);
+
+    /** Used by tests only.
+    */
+    std::size_t
+    setMinimumTx(int m)
+    {
+        std::lock_guard <std::mutex> sl(lock_);
+
+        auto const old = minimumTxnCount_;
+        minimumTxnCount_ = m;
+        txnsExpected_ = m;
+        return old;
+    }
+
+    std::size_t
+    getTxnsExpected() const
+    {
+        std::lock_guard <std::mutex> sl(lock_);
+
+        return txnsExpected_;
+    }
+
+    std::uint32_t
+    getEscalationMultiplier() const
+    {
+        std::lock_guard <std::mutex> sl(lock_);
+
+        return escalationMultiplier_;
+    }
+
+    std::uint64_t
+    scaleFeeLevel(OpenView const& view) const;
+};
+
+}
 
 /**
     Transaction Queue. Used to manage transactions in conjunction with
@@ -46,9 +134,6 @@ class Application;
     are first applied from the queue to the open ledger in fee order
     until either all transactions are applied or the fee again jumps
     too high for the remaining transactions.
-
-    Note EAHENNIS: I hate that this class is virtualized, but
-    flattening it out introduces dependency problems.
 */
 class TxQ
 {
@@ -72,7 +157,10 @@ public:
         std::uint64_t expFeeLevel;        // Estimated fee level to get in next ledger
     };
 
-    virtual ~TxQ() = default;
+    TxQ(Setup const& setup,
+        beast::Journal j);
+
+    virtual ~TxQ();
 
     /**
         Add a new transaction to the open ledger, hold it in the queue,
@@ -105,11 +193,10 @@ public:
         @param params Flags to control engine behaviors.
         @param engine Transaction Engine.
     */
-    virtual
     std::pair<TER, bool>
     apply(Application& app, OpenView& view,
         std::shared_ptr<STTx const> const& tx,
-            ApplyFlags flags, beast::Journal j) = 0;
+            ApplyFlags flags, beast::Journal j);
 
     /**
         Fill the new open ledger with transactions from the queue.
@@ -126,10 +213,9 @@ public:
 
         @return Whether any txs were added to the view.
     */
-    virtual
     bool
     accept(Application& app, OpenView& view,
-        ApplyFlags flags = tapNONE) = 0;
+        ApplyFlags flags = tapNONE);
 
     /**
         We have a new last validated ledger, update and clean up the
@@ -145,36 +231,138 @@ public:
             `LastLedgerSequence` has passed.
 
     */
-    virtual
     void
     processValidatedLedger(Application& app,
         OpenView const& view, bool timeLeap,
-            ApplyFlags flags = tapNONE) = 0;
+            ApplyFlags flags = tapNONE);
 
     /** Used by tests only.
     */
-    virtual
     std::size_t
-    setMinimumTx(int m) = 0;
+    setMinimumTx(int m);
 
     /** Returns fee metrics in reference fee (level) units.
     */
-    virtual
     struct Metrics
-    getMetrics(OpenView const& view) const = 0;
+    getMetrics(OpenView const& view) const;
 
     /** Packages up fee metrics for the `fee` RPC command.
     */
-    virtual
     Json::Value
-    doRPC(Application& app) const = 0;
+    doRPC(Application& app) const;
 
     /** Return the instantaneous fee to get into the current
         open ledger for a reference transaction.
     */
-    virtual
     XRPAmount
-    openLedgerFee(OpenView const& view) const = 0;
+    openLedgerFee(OpenView const& view) const;
+
+private:
+    class CandidateTxn
+    {
+    public:
+        // Used by the TxQ::FeeHook and TxQ::FeeMultiSet below
+        // to put each candidate object into more than one
+        // set without copies, pointers, etc.
+        boost::intrusive::set_member_hook<> byFeeListHook;
+
+        std::shared_ptr<STTx const> txn;
+
+        uint64_t const feeLevel;
+        TxID const txID;
+        boost::optional<TxID> priorTxID;
+        AccountID const account;
+        boost::optional<LedgerIndex> lastValid;
+        TxSeq const sequence;
+        ApplyFlags const flags;
+        // pfresult_ is never allowed to be empty. The
+        // boost::optional is leveraged to allow `emplace`d
+        // construction and replacement without a copy
+        // assignment operation.
+        boost::optional<PreflightResult const> pfresult;
+
+    public:
+        CandidateTxn(std::shared_ptr<STTx const> const&,
+            TxID const& txID, std::uint64_t feeLevel,
+                ApplyFlags const flags,
+                    PreflightResult const& pfresult);
+    };
+
+    class GreaterFee
+    {
+    public:
+        bool operator()(const CandidateTxn& lhs, const CandidateTxn& rhs) const
+        {
+            return lhs.feeLevel > rhs.feeLevel;
+        }
+    };
+
+    class TxQAccount
+    {
+    public:
+
+        AccountID const account;
+        uint64_t totalFees;
+        // Sequence number will be used as the key.
+        std::map <TxSeq, CandidateTxn> transactions;
+
+    public:
+        explicit TxQAccount(std::shared_ptr<STTx const> const& txn);
+        explicit TxQAccount(const AccountID& account);
+
+        std::size_t
+        getTxnCount() const
+        {
+            return transactions.size();
+        }
+
+        bool
+        empty() const
+        {
+            return !getTxnCount();
+        }
+
+        CandidateTxn&
+        addCandidate(CandidateTxn&&);
+
+        bool
+        removeCandidate(TxSeq const& sequence);
+
+        CandidateTxn const*
+        findCandidateAt(TxSeq const& sequence) const;
+    };
+
+
+    using FeeHook = boost::intrusive::member_hook
+        <CandidateTxn, boost::intrusive::set_member_hook<>,
+        &CandidateTxn::byFeeListHook>;
+
+    using FeeMultiSet = boost::intrusive::multiset
+        < CandidateTxn, FeeHook,
+        boost::intrusive::compare <GreaterFee> >;
+
+    Setup const setup_;
+    beast::Journal j_;
+
+    detail::FeeMetrics feeMetrics_;
+    FeeMultiSet byFee_;
+    std::map <AccountID, TxQAccount> byAccount_;
+    boost::optional<size_t> maxSize_;
+
+    // Most queue operations are done under the master lock,
+    // but use this mutex for the RPC "fee" command, which isn't.
+    std::mutex mutable mutex_;
+
+private:
+    bool isFull() const
+    {
+        return maxSize_ && byFee_.size() >= *maxSize_;
+    }
+
+    bool canBeHeld(std::shared_ptr<STTx const> const&);
+
+    FeeMultiSet::iterator_type erase(FeeMultiSet::const_iterator_type);
+
 };
 
 TxQ::Setup
