+++ src/ripple/ledger/impl/ReadView.cpp
@@ -110,7 +110,7 @@ bool
 Rules::operator== (Rules const& other) const
 {
 #if 1
-    if (! impl_ && ! other.impl_)
+    if (impl_.get() == other.impl_.get())
         return true;
     if (! impl_ || ! other.impl_)
         return false;
