+++ src/ripple/unity/app_misc.cpp
@@ -29,3 +29,4 @@
 #include <ripple/app/misc/Validations.cpp>
 
 #include <ripple/app/misc/impl/AccountTxPaging.cpp>
+#include <ripple/app/misc/impl/TxQ.cpp>
