+++ src/ripple/unity/app_tx.cpp
@@ -20,6 +20,7 @@
 #include <BeastConfig.h>
 
 #include <ripple/app/tx/impl/apply.cpp>
+#include <ripple/app/tx/impl/applySteps.cpp>
 #include <ripple/app/tx/impl/BookTip.cpp>
 #include <ripple/app/tx/impl/CancelOffer.cpp>
 #include <ripple/app/tx/impl/CancelTicket.cpp>
@@ -42,4 +43,3 @@
 #include <ripple/app/tx/impl/ApplyContext.cpp>
 #include <ripple/app/tx/impl/TransactionAcquire.cpp>
 #include <ripple/app/tx/impl/Transactor.cpp>
-#include <ripple/app/tx/impl/TxQImpl.cpp>
